Create Python_task1.py
Hello everyone,
My name is Rohit Kumar Patra
this is a python code for a mastermind game which was given as a task by Robotics club, VSSUT, Burla 
*Rules of the game* (2 Player Version):
1. One player writes down a secret 4-digit number.
2. Then, the other player tries to guess the number.
3. After each guess, the person that has written down the secret
number, tells them how many digits are correct and in the right
position and how many digits are correct but in the wrong place.
4. Players keep making guesses until they are able to guess the 4 digit
number correctly.
5. There can be a limit the number of guesses the 2nd player can have.

1. The program should generate a random 4 digit number at the start
2. The player starts with 20 points
3. On each turn, the program asks for a number from the user.
4. After taking input, the program displays the digits that are correct and
in the right position and how many digits are correct but in the wrong
place. Also display the which digits are in correct position
5. Program ends when The numbers of guesses reaches a number (Say 10) or The player guesses the number correctly
6. For each correct guess +5 points are given and for each incorrect
guess -2 points are given.
7. On completion, Score for the user should be displayed and options to
play again should be provided.